requirements.txt
# FastAPI backend
fastapi
uvicorn[standard]

# OpenAI API (GPT-4o + Whisper)
openai

# ElevenLabs TTS
requests

# Audio recording / playback (for local CLI testing)
sounddevice
soundfile
pygame

# Env management
python-dotenv

# Twilio voice & SMS
twilio

# Optional CORS handling
aiofiles
python-multipart
sounddevice
pygame
requests
python-dotenv
